+++ crates/wasmtime/src/runtime/linker.rs
@@ -5,15 +5,13 @@ use crate::store::StoreOpaque;
 use crate::{prelude::*, IntoFunc};
 use crate::{
     AsContext, AsContextMut, Caller, Engine, Extern, ExternType, Func, FuncType, ImportType,
-    Instance, Module, StoreContextMut, Val, ValRaw, ValType, WasmTyList,
+    Instance, Module, StoreContextMut, Val, ValRaw, ValType,
 };
 use alloc::sync::Arc;
 use core::fmt;
-#[cfg(feature = "async")]
-use core::future::Future;
 use core::marker;
 #[cfg(feature = "async")]
-use core::pin::Pin;
+use core::{future::Future, pin::Pin};
 use log::warn;
 
 /// Structure used to link wasm modules/instances together.
@@ -465,8 +463,8 @@ impl<T> Linker<T> {
                 .0
                 .async_cx()
                 .expect("Attempt to spawn new function on dying fiber");
-            let mut future = Pin::from(func(caller, params, results));
-            match unsafe { async_cx.block_on(future.as_mut()) } {
+            let future = func(caller, params, results);
+            match unsafe { async_cx.block_on(Pin::from(future)) } {
                 Ok(Ok(())) => Ok(()),
                 Ok(Err(trap)) | Err(trap) => Err(trap),
             }
@@ -543,7 +541,7 @@ impl<T> Linker<T> {
 
     /// Asynchronous analog of [`Linker::func_wrap`].
     #[cfg(feature = "async")]
-    pub fn func_wrap_async<F, Params: WasmTyList, Args: crate::WasmRet>(
+    pub fn func_wrap_async<F, Params: crate::WasmTyList, Args: crate::WasmRet>(
         &mut self,
         module: &str,
         name: &str,
@@ -568,8 +566,8 @@ impl<T> Linker<T> {
                     .0
                     .async_cx()
                     .expect("Attempt to start async function on dying fiber");
-                let mut future = Pin::from(func(caller, args));
-                match unsafe { async_cx.block_on(future.as_mut()) } {
+                let future = func(caller, args);
+                match unsafe { async_cx.block_on(Pin::from(future)) } {
                     Ok(ret) => ret.into_fallible(),
                     Err(e) => Args::fallible_from_error(e),
                 }
